Refactor code to pass only configuration options to child processes
Moved calculation of submats to within the child processes and brought in functions
required for entire update. The new structure requries that the matrix be a saved .npz file
and be in scipy.sparse.csr_matrix format. The required command line arguments for running
the codea are rank, tolerance, and regularization penalty.